+++ BurningKnight/entity/events/PostHealthModifiedEvent.cs
@@ -2,7 +2,7 @@ using Lens.entity;
 
 namespace BurningKnight.entity.events {
 	public class PostHealthModifiedEvent : Event {
-		public int Amount;
+		public float Amount;
 		public Entity From;
 		public Entity Who;
 		public bool Default = true;
